login and product endpoint changes

+++ routes/products.js
@@ -6,14 +6,35 @@ const jwt = require('jsonwebtoken');
 Router.get("/", async (req, res) => {
     let query = "SELECT * FROM products";
     let result = await sequelize.query(query, { raw: true });
+
+    // const secreto = 'AlgunSecret0';
+    // const token = jwt.sign({
+    //     nombre: 'Sergio',
+    // }, secreto);
+    
+    // Imprimir token generado
+    // console.log(token);
+
     res.send(result[0]);
+    // res.send(token);
 })
 
 Router.get("/:productId", async (req, res) => {
     const productId = req.params.productId;
     let query = `SELECT * FROM products WHERE product_id = ${productId}`;
-    let result = await sequelize.query(query, { raw: true });
-    res.send(result[0]);
+    try{
+        let result = await sequelize.query(query, { raw: true });
+        if(result[0].length > 0) {
+            res.send(result[0]);
+        } else {
+            res.status(404);
+            res.send("No se ha encontrado un producto con el ID especificado")
+        }
+    }catch{
+        res.status(500);
+        res.send("Ha ocurrido un error en el servidor")
+    }
+    
 })
 
 Router.post("/", async (req, res) => {
@@ -33,7 +54,7 @@ Router.put("/", async (req, res) => {
     let query = `UPDATE products SET title = '${req.body.title}', price = '${req.body.price}' WHERE product_id = ${req.body.product_id}`;
     try{
         let result = await sequelize.query(query, { raw: true });
-        res.status(res.statusCode);
+        res.status(200);
         res.send(req.body);
     }catch(err){
         res.status(res.statusCode);
@@ -42,15 +63,23 @@ Router.put("/", async (req, res) => {
 })
 
 Router.delete("/:productId", async (req, res) => {
-    let query = `DELETE FROM products WHERE product_id = ${req.params.productId}`
-    try{
-        let result = await sequelize.query(query, { raw: true });
-        res.status(res.statusCode);
-        res.send("El producto ha sido eliminado");
-    }catch(err){
-        res.status(res.statusCode);
-        res.send("Ha ocurrido un error al eliminar el producto")
+    let query = `SELECT * FROM products WHERE product_id = ${req.params.productId}`
+    let result = await sequelize.query(query, { raw: true });
+    if ( result[0].length > 0 ) {
+        let query = `DELETE FROM products WHERE product_id = ${req.params.productId}`
+        try{
+            let result = await sequelize.query(query, { raw: true });
+            res.status(200);
+            res.send("El producto ha sido eliminado");
+        }catch(err){
+            res.status(500);
+            res.send("Ha ocurrido un error al eliminar el producto")
+        }
+    } else {
+        res.status(404);
+        res.send("El ID indicado no corresponde a un producto existente")
     }
+    
 })
 
 module.exports = Router;
